uploads images

+++ lib/core/widgets/my_app_bar.dart
@@ -5,42 +5,37 @@ import 'package:cloud_medix/core/theming/styles.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
-import 'package:flutter_svg/flutter_svg.dart';
 
 class MyAppBar extends StatelessWidget implements PreferredSizeWidget {
-  MyAppBar({
+  const MyAppBar({
     super.key,
     required this.title,
     this.isSettings = false, // optional with default
   });
 
   final String title;
-  bool isSettings = false;
+  final bool isSettings;
   @override
   Widget build(BuildContext context) {
     return AppBar(
       backgroundColor: ColorsManager.primaryColor,
       actions: [
-        isSettings
-            ? IconButton(
-                icon: const Icon(
-                  Icons.logout,
-                  color: Colors.white,
-                ),
-                onPressed: () async {
-                  FlutterSecureStorage storage = getIt<FlutterSecureStorage>();
-                  await storage.deleteAll();
-                  if (context.mounted) {
-                    WidgetsBinding.instance.addPostFrameCallback((_) {
-                      Navigator.popAndPushNamed(context, Routes.login);
-                    });
-                  }
-                },
-              )
-            : Padding(
-                padding: EdgeInsets.symmetric(vertical: 3.h),
-                child: SvgPicture.asset('assets/images/user.svg'),
-              ),
+        if (isSettings)
+          IconButton(
+            icon: const Icon(
+              Icons.logout,
+              color: Colors.white,
+            ),
+            onPressed: () async {
+              FlutterSecureStorage storage = getIt<FlutterSecureStorage>();
+              await storage.deleteAll();
+              if (context.mounted) {
+                WidgetsBinding.instance.addPostFrameCallback((_) {
+                  Navigator.popAndPushNamed(context, Routes.login);
+                });
+              }
+            },
+          )
       ],
       title: Text(title),
       titleTextStyle: TextStyles.appBarTexts,
